Changed GPIO Off timeout to long, as int is too short

+++ JeePower2.ino
@@ -49,7 +49,7 @@ int OilPressureOffTimeout = 30000;	// Timeout before confirming oil pressure war
 long IgnitionOnMillis = 0;				// Time the ignition came on
 long IgnitionOffMillis = 0;			// Time the ignition went off
 long OilPressureOffMillis = 0;		// time the oil pressure warning went off
-int GPIOOffTimeout = 300000;		// This is the time between turning off the GPIO relay, and turning off the main relay (10 minutes)
+long GPIOOffTimeout = 300000;		// This is the time between turning off the GPIO relay, and turning off the main relay (10 minutes)
 
 long GPIOOffTime = 0;					// Time the GPIO relay is disabled
 
@@ -156,7 +156,7 @@ void loop(){
 	 if (OilState == 0 && OilPressureOffMillis == 0) {
 		  // The oil light is off. We're good to go
 		  if (DEBUG) { Serial.println("Oil pressure ok."); }
-		  OilPressureOffMillis = CurrentMillis;
+		          OilPressureOffMillis = CurrentMillis;
 		  delay(1000);
 	 }
